[ig-colors] Move asymptote interaction stories to interactions regression stories
